requirements.txt
langchain-google-genai 
crewai 
langchain-community 
crewai-tools 
python-dotenv
duckduckgo-search